Moved credmark.types.dto to credmark.dto
