+++ models/credmark/algorithms/var.py
@@ -14,7 +14,7 @@ from datetime import (
 import credmark.model
 from credmark.model import ModelRunError
 
-from credmark.types.dto import (
+from credmark.dto import (
     DTO,
     DTOField,
 )
