+++ models/examples/dto_examples.py
@@ -1,5 +1,5 @@
 import credmark.model
-from credmark.types.dto import DTO, DTOField
+from credmark.dto import DTO, DTOField
 from credmark.types import Portfolio
