+++ models/examples/echo_example.py
@@ -1,5 +1,5 @@
 import credmark.model
-from credmark.types.dto import DTO, DTOField
+from credmark.dto import DTO, DTOField
 
 
 class EchoDto(DTO):
