+++ models/examples/historical_examples.py
@@ -1,5 +1,5 @@
 import credmark.model
-from credmark.types.dto import DTO
+from credmark.dto import DTO
 from models.examples.library_examples import LibrariesDto
